Move gap_report utils into gap_report module
