+++ deafrica/monitoring/s2_c1_gap_filler.py
@@ -16,9 +16,11 @@ from rasterio.session import AWSSession
 from deafrica import __version__
 from deafrica.click_options import slack_url
 from deafrica.logs import setup_logging
-from deafrica.utils import (
+from deafrica.monitoring.gap_report import (
     find_latest_report,
     read_report_missing_scenes,
+)
+from deafrica.utils import (
     send_slack_notification,
     split_list_equally,
 )
