+++ deafrica/monitoring/s2_gap_filler.py
@@ -19,9 +19,11 @@ from shapely import geometry
 from deafrica import __version__
 from deafrica.click_options import limit, slack_url
 from deafrica.logs import setup_logging
-from deafrica.utils import (
+from deafrica.monitoring.gap_report import (
     find_latest_report,
     read_report_missing_scenes,
+)
+from deafrica.utils import (
     send_slack_notification,
     split_list_equally,
 )
